refactor: group analysis and config modules
