+++ analyze_git_repo_loc/__main__.py
@@ -32,7 +32,7 @@ import plotly.graph_objects as go
 from colorama import Cursor, Fore, Style
 from tqdm import tqdm
 
-from analyze_git_repo_loc.analysis_helpers import (
+from analyze_git_repo_loc.analysis.analysis_helpers import (
     prepare_author_contribution_data,
     prepare_summary_data,
     prepare_trend_data,
@@ -45,7 +45,7 @@ from analyze_git_repo_loc.reporting.html_report import (
     generate_html_report,
 )
 from analyze_git_repo_loc.i18n import tr
-from analyze_git_repo_loc.init_wizard import run_init_config_wizard
+from analyze_git_repo_loc.config.init_wizard import run_init_config_wizard
 from analyze_git_repo_loc.reporting.markdown_summary import generate_markdown_summary
 from analyze_git_repo_loc.remote.remote_catalog import (
     RemoteCatalogError,
@@ -60,7 +60,7 @@ from analyze_git_repo_loc.utils import (
     parse_arguments,
     save_repository_branch_info,
 )
-from analyze_git_repo_loc.yaml_config import load_yaml_data
+from analyze_git_repo_loc.config.yaml_config import load_yaml_data
 
 
 class _ReportProgressTracker:
