+++ analyze_git_repo_loc/analysis/__init__.py
@@ -1,0 +1,6 @@
+"""Analysis pipeline package.
+
+Description:
+    Groups repository analysis helpers, exclude-template logic, and analyzer
+    runtime components used by CLI and interactive execution.
+"""
