+++ analyze_git_repo_loc/config/__init__.py
@@ -1,0 +1,6 @@
+"""Configuration workflow package.
+
+Description:
+    Groups YAML loading and init-wizard configuration builders used by batch
+    and interactive setup flows.
+"""
